Add files via upload
The program I wrote is called "LawsByAge" and it is designed to identify and derive laws that apply to a person based on their age. Here's how this program works:

The user runs the program and enters their age through the console interface.

The program checks the entered age and determines whether it is less than 18 years old.

If the age is less than 18 years, the program displays a list of laws and regulations that apply to persons under 18 years of age. Laws in this category contain restrictions and regulations that protect the rights and safety of minors.

If the age is equal to or greater than 18 years, the program displays a list of laws and rights that apply to adults. These laws may include rights and responsibilities related to elections, work, education, property, and other aspects of adult life.

After deriving the necessary laws, the program completes its work.

The "LawsByAge" program provides an opportunity for the user to obtain general information about the laws that may affect their life depending on age. It can be useful for educating and informing citizens about their rights and responsibilities in different periods of life.